refactor(Tooltip): client/配下をフラット配置に整理する
client/components/Tooltip/ という二重ネストを解消し、client/ 直下に
フラットに配置する。Tooltip配下にはhookが存在せず、components/という
下位区分自体に意味がなかったため。

判断基準はCLAUDE.mdの「client 境界が必要なモジュールは client/ に
閉じ込める」節を参照。
