--- packages/smarthr-ui/src/components/Tooltip/client/components/Tooltip/index.ts
@@ -1,1 +1,0 @@
-export { Tooltip } from './Tooltip'
