+++ packages/smarthr-ui/src/components/Tooltip/index.ts
@@ -1,2 +1,2 @@
 export { ControlledTooltip } from './ControlledTooltip'
-export { Tooltip } from './client/components'
+export { Tooltip } from './client'
