+++ packages/smarthr-ui/src/components/Tooltip/stories/Tooltip.stories.tsx
@@ -1,7 +1,7 @@
 import { Button } from '../../Button'
 import { FaCircleQuestionIcon, FaPencilIcon } from '../../Icon'
 import { Cluster, Stack } from '../../Layout'
-import { Tooltip } from '../client/components'
+import { Tooltip } from '../client'
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
